Add a summary document for the dataframe interchange protocol
Summarizes the various discussions about and goals/non-goals
and requirements for the `__dataframe__` data interchange
protocol.

The intended audience for this document is Consortium members
and dataframe library maintainers who may want to support this
protocol. @datapythonista will add a companion document that's
a more gentle introduction/tutorial in a "from zero to a protocol"
style.

The aim is to keep updating this till we have captured all
the requirements and answered all the FAQs, so we can actually
design the protocol after and verify it meets all our requirements.

Closes gh-29