Illustrate issue with categorical dtype & get_data_buffer()
This shows the simple design doesn't fully work (see the FIXMEs
in the diff). Instead, the `children` concept is needed.
That way the categorical encoded data values can be returned
as a child Column rather than a Buffer, and hence there's the
necessary Column.dtype to interpret the buffer backing the column.